Move `DummyTest` to `Stub` sub-namespace
